# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SplitTab – bill-splitting API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Group {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  currency: string;
  createdAt: string;
}

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  currency?: string;
}

export interface GroupUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  currency?: string;
}

export interface Member {
  id: number;
  groupId: number;
  name: string;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  email?: string;
}

export type ExpenseSplitMode = typeof ExpenseSplitMode[keyof typeof ExpenseSplitMode];


export const ExpenseSplitMode = {
  equal: 'equal',
  amount: 'amount',
  shares: 'shares',
} as const;

export interface Expense {
  id: number;
  groupId: number;
  title: string;
  amount: number;
  currency: string;
  /** ISO 8601 date (YYYY-MM-DD) */
  date: string;
  payerId: number;
  payerName?: string;
  /**
     * Optional product size or quantity descriptor (e.g. "2L", "Family pack")
     * @nullable
     */
  productSize?: string | null;
  /** @nullable */
  notes?: string | null;
  splitMode: ExpenseSplitMode;
  createdAt: string;
}

export type ExpenseInputSplitMode = typeof ExpenseInputSplitMode[keyof typeof ExpenseInputSplitMode];


export const ExpenseInputSplitMode = {
  equal: 'equal',
  amount: 'amount',
  shares: 'shares',
} as const;

export interface SplitInput {
  memberId: number;
  owedAmount?: number;
  shares?: number;
}

export interface ExpenseInput {
  /** @minLength 1 */
  title: string;
  amount: number;
  currency?: string;
  date: string;
  payerId: number;
  productSize?: string;
  notes?: string;
  splitMode: ExpenseInputSplitMode;
  splits: SplitInput[];
}

export type ExpenseUpdateSplitMode = typeof ExpenseUpdateSplitMode[keyof typeof ExpenseUpdateSplitMode];


export const ExpenseUpdateSplitMode = {
  equal: 'equal',
  amount: 'amount',
  shares: 'shares',
} as const;

export interface ExpenseUpdate {
  /** @minLength 1 */
  title?: string;
  amount?: number;
  currency?: string;
  date?: string;
  payerId?: number;
  productSize?: string;
  notes?: string;
  splitMode?: ExpenseUpdateSplitMode;
  splits?: SplitInput[];
}

export type ExpenseDetailSplitMode = typeof ExpenseDetailSplitMode[keyof typeof ExpenseDetailSplitMode];


export const ExpenseDetailSplitMode = {
  equal: 'equal',
  amount: 'amount',
  shares: 'shares',
} as const;

export interface Split {
  id: number;
  expenseId: number;
  memberId: number;
  memberName: string;
  owedAmount: number;
  /** @nullable */
  shares?: number | null;
}

export interface ExpenseDetail {
  id: number;
  groupId: number;
  title: string;
  amount: number;
  currency: string;
  date: string;
  payerId: number;
  payerName?: string;
  /** @nullable */
  productSize?: string | null;
  /** @nullable */
  notes?: string | null;
  splitMode: ExpenseDetailSplitMode;
  splits: Split[];
  createdAt: string;
}

export interface Payment {
  id: number;
  groupId: number;
  fromMemberId: number;
  fromMemberName?: string;
  toMemberId: number;
  toMemberName?: string;
  amount: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PaymentInput {
  fromMemberId: number;
  toMemberId: number;
  amount: number;
  date: string;
  notes?: string;
}

export interface MemberBalance {
  memberId: number;
  memberName: string;
  /** Positive = is owed money; negative = owes money */
  netBalance: number;
}

export interface Settlement {
  fromMemberId: number;
  fromMemberName: string;
  toMemberId: number;
  toMemberName: string;
  amount: number;
}

export interface GroupSummary {
  groupId: number;
  totalSpent: number;
  memberCount: number;
  expenseCount: number;
  currency: string;
  recentExpenses: Expense[];
}