# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SplitTab – bill-splitting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups/{groupId}/expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /groups/{groupId}/expenses/{expenseId}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense with its splits
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
        - name: expenseId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Expense with splits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
        - name: expenseId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
        - name: expenseId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ── Payments ─────────────────────────────────────────────────────────────────
...
  /groups/{groupId}/balances:
    get:
      operationId: getBalances
      tags: [balances]
      summary: Get net balances for each member
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Member balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberBalance"
  /groups/{groupId}/settlements:
    get:
      operationId: getSettlements
      tags: [balances]
      summary: Get suggested payments to settle all debts
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Settlement suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Settlement"
  /groups/{groupId}/summary:
    get:
      operationId: getGroupSummary
      tags: [balances]
      summary: Group overview – totals, member count, recent activity
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupSummary"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Group ──────────────────────────────────────────────────────────────────
...
    Expense:
      type: object
      required: [id, groupId, title, amount, currency, date, payerId, splitMode, createdAt]
      properties:
        id:
          type: number
        groupId:
          type: number
        title:
          type: string
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
          description: ISO 8601 date (YYYY-MM-DD)
        payerId:
          type: number
        payerName:
          type: string
        productSize:
          type: ["string", "null"]
          description: Optional product size or quantity descriptor (e.g. "2L", "Family pack")
        notes:
          type: ["string", "null"]
        splitMode:
          type: string
          enum: [equal, amount, shares]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [title, amount, date, payerId, splitMode, splits]
      properties:
        title:
          type: string
          minLength: 1
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        payerId:
          type: number
        productSize:
          type: string
        notes:
          type: string
        splitMode:
          type: string
          enum: [equal, amount, shares]
        splits:
          type: array
          items:
            $ref: "#/components/schemas/SplitInput"
    ExpenseUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        payerId:
          type: number
        productSize:
          type: string
        notes:
          type: string
        splitMode:
          type: string
          enum: [equal, amount, shares]
        splits:
          type: array
          items:
            $ref: "#/components/schemas/SplitInput"
    ExpenseDetail:
      type: object
      required: [id, groupId, title, amount, currency, date, payerId, splitMode, splits, createdAt]
      properties:
        id:
          type: number
        groupId:
          type: number
        title:
          type: string
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        payerId:
          type: number
        payerName:
          type: string
        productSize:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        splitMode:
          type: string
          enum: [equal, amount, shares]
        splits:
          type: array
          items:
            $ref: "#/components/schemas/Split"
        createdAt:
          type: string
    Split:
      type: object
      required: [id, expenseId, memberId, memberName, owedAmount]
      properties:
        id:
          type: number
        expenseId:
          type: number
        memberId:
          type: number
        memberName:
          type: string
        owedAmount:
          type: number
        shares:
          type: ["number", "null"]
    SplitInput:
      type: object
      required: [memberId]
      properties:
        memberId:
          type: number
        owedAmount:
          type: number
        shares:
          type: number

    # ── Payment ────────────────────────────────────────────────────────────────
...
    MemberBalance:
      type: object
      required: [memberId, memberName, netBalance]
      properties:
        memberId:
          type: number
        memberName:
          type: string
        netBalance:
          type: number
          description: Positive = is owed money; negative = owes money
    Settlement:
      type: object
      required: [fromMemberId, fromMemberName, toMemberId, toMemberName, amount]
      properties:
        fromMemberId:
          type: number
        fromMemberName:
          type: string
        toMemberId:
          type: number
        toMemberName:
          type: string
        amount:
          type: number
    GroupSummary:
      type: object
      required: [groupId, totalSpent, memberCount, expenseCount, currency, recentExpenses]
      properties:
        groupId:
          type: number
        totalSpent:
          type: number
        memberCount:
          type: number
        expenseCount:
          type: number
        currency:
          type: string
        recentExpenses:
          type: array
          items:
            $ref: "#/components/schemas/Expense"