# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SplitTab – bill-splitting API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseSplitMode } from './expenseSplitMode';

export interface Expense {
  id: number;
  groupId: number;
  title: string;
  amount: number;
  currency: string;
  /** ISO 8601 date (YYYY-MM-DD) */
  date: string;
  payerId: number;
  payerName?: string;
  /**
     * Optional product size or quantity descriptor (e.g. "2L", "Family pack")
     * @nullable
     */
  productSize?: string | null;
  /** @nullable */
  notes?: string | null;
  splitMode: ExpenseSplitMode;
  createdAt: string;
}