# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SplitTab – bill-splitting API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseInputSplitMode } from './expenseInputSplitMode';
import type { SplitInput } from './splitInput';

export interface ExpenseInput {
  /** @minLength 1 */
  title: string;
  amount: number;
  currency?: string;
  date: string;
  payerId: number;
  productSize?: string;
  notes?: string;
  splitMode: ExpenseInputSplitMode;
  splits: SplitInput[];
}