# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SplitTab – bill-splitting API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseUpdateSplitMode } from './expenseUpdateSplitMode';
import type { SplitInput } from './splitInput';

export interface ExpenseUpdate {
  /** @minLength 1 */
  title?: string;
  amount?: number;
  currency?: string;
  date?: string;
  payerId?: number;
  productSize?: string;
  notes?: string;
  splitMode?: ExpenseUpdateSplitMode;
  splits?: SplitInput[];
}