# lib/api-zod/src/generated/types/groupSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SplitTab – bill-splitting API
 * OpenAPI spec version: 0.1.0
 */
import type { Expense } from './expense';

export interface GroupSummary {
  groupId: number;
  totalSpent: number;
  memberCount: number;
  expenseCount: number;
  currency: string;
  recentExpenses: Expense[];
}