# lib/api-zod/src/generated/types/splitInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SplitTab – bill-splitting API
 * OpenAPI spec version: 0.1.0
 */

export interface SplitInput {
  memberId: number;
  owedAmount?: number;
  shares?: number;
}